index/function-source-V-1/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud
google-cloud-storage
pandas
openpyxl
google-cloud-datastore